Move graphs to graph folder
