+++ job_visualization/graphs/call_graph.py
@@ -2,10 +2,10 @@
 import graphviz as gv
 
 import collections
-from file_data import FileData
+from ..file_data import FileData
 
-from graph import Graph
-import yaml_unfolder
+from .graph import Graph
+from ..util import convert_path
 
 CallEdge = collections.namedtuple('CallEdge', ['project_name', 'call_config'])
 
@@ -231,4 +231,4 @@ class CallGraph(Graph):
 
     def get_path_from_name(self, name):
         path = self._configs[name].path
-        return yaml_unfolder.convert_path(path)
+        return convert_path(path)
