+++ job_visualization/graphs/legend.py
