+++ job_visualization/test/test_syn_parser/test_syn_parser.py
@@ -6,7 +6,7 @@ import os
 sys.path.append("../")
 sys.path.append("../../")
 
-from synonym_parser import *
+from job_visualization.synonym_parser import *
 
 class TestSynonymParser(unittest.TestCase):
